# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CostManagement Api Documentation",
        "description": "Implementation of CostManagement Api",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/ExpansiveCost/addExpansiveCost": {
            "post": {
                "tags": [
                    "ExpansiveCosts"
                ],
                "summary": "add for new ExpansiveCost ",
                "description": "",
                "operationId": "1-addExpansiveCost",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "itemCode": {
                                        "type": "string",
                                        "example": "10k1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "A water bill"
                                    },
                                    "paidValue": {
                                        "type": "number",
                                        "example": 20200
                                    },
                                    "date": {
                                        "type": "date",
                                        "example": "2020/01/20"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/ExpansiveCost/editExpansiveCost": {
            "put": {
                "tags": [
                    "ExpansiveCosts"
                ],
                "summary": "edit for new ExpansiveCost ",
                "description": "",
                "operationId": "2-editExpansiveCost",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "number",
                                        "example": 1
                                    },
                                    "itemCode": {
                                        "type": "string",
                                        "example": "10k1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "A water bill"
                                    },
                                    "paidValue": {
                                        "type": "number",
                                        "example": 20200
                                    },
                                    "date": {
                                        "type": "date",
                                        "example": "2020/01/20"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/ExpansiveCost/deleteExpansiveCost/{id}": {
            "delete": {
                "tags": [
                    "ExpansiveCosts"
                ],
                "summary": "delet from ExpansiveCosts by id",
                "description": "",
                "operationId": "3-deleteExpansiveCostQuestion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/ExpansiveCost/showExpansiveCostById/{id}": {
            "get": {
                "tags": [
                    "ExpansiveCosts"
                ],
                "summary": "show from ExpansiveCost By ID",
                "description": "Returns info spcific ExpansiveCost",
                "operationId": "4-showExpansiveCostById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/ExpansiveCost/showItemsByExpansiveCode/{code}": {
            "get": {
                "tags": [
                    "ExpansiveCosts"
                ],
                "summary": "show list of ItemCost",
                "description": "Returns list of ItemCost",
                "operationId": "5-showItemsByExpansiveCode",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/ExpansiveCost/showAllExpansiveCost": {
            "get": {
                "tags": [
                    "ExpansiveCosts"
                ],
                "summary": "show list of ExpansiveCost",
                "description": "Returns list of all ExpansiveCost",
                "operationId": "6-showAllExpansiveCost",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/ExpansiveCost/changeState/{id}/{state}": {
            "put": {
                "tags": [
                    "ExpansiveCosts"
                ],
                "summary": "changeState from ExpansiveCost",
                "description": "change state  ExpansiveCost to any state (1-paid or 2-unpaid)",
                "operationId": "7-changeStateExpansiveCost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "state",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    }
}